requirements.txt
absl-py==1.4.0
accelerate==1.2.1
aiofiles==23.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
altair==5.0.1
altgraph==0.17.4
annotated-types==0.7.0
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
APScheduler==3.11.0
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.45.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
blis==1.1.0
cachetools==5.3.0
catalogue==2.0.10
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.20.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.3
confection==0.1.5
contourpy==1.0.7
cryptography==43.0.1
cssselect2==0.7.0
cvzone==1.6.1
cycler==0.11.0
cymem==2.0.10
Cython==3.0.11
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
demucs==4.0.1
diff-match-patch==20230430
distro==1.9.0
Django==5.0.2
django-admin-interface==0.28.5
django-colorfield==0.11.0
django-htmx==1.17.3
django-import-export==3.3.7
django-multiselectfield==0.1.12
django-widget-tweaks==1.5.0
dlib==19.24.2
dm-tree==0.1.8
dnspython==1.16.0
docker-pycreds==0.4.0
docopt==0.6.2
dora_search==0.1.12
easyocr==1.7.2
einops==0.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==1.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.115.6
fastjsonschema==2.16.3
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.12.0
fire==0.7.0
Flask==2.3.2
flatbuffers==23.5.26
flyingcircus==0.1.4.1
fonttools==4.39.3
fpdf==1.7.2
fqdn==1.5.1
freetype-py==2.3.0
frozendict==2.4.0
frozenlist==1.3.3
fsspec==2023.5.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.4.0
gensim==4.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==5.12.0
gradio_client==1.5.4
grpcio==1.54.0
gTTS==2.5.4
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.25.1
idna==3.4
imageio==2.36.1
imgaug==0.4.0
inflect==7.0.0
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.9
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.1
jiter==0.6.1
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.1
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.10
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.25.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lameenc==1.7.0
langcodes==3.5.0
language-tool-python==2.7.1
language_data==1.3.0
lazy_loader==0.4
libclang==16.0.0
linkify-it-py==2.0.2
llvmlite==0.40.0
lmdb==1.6.2
lxml==4.9.2
marisa-trie==1.2.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.11
mistune==2.0.5
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.4
multitasking==0.0.11
murmurhash==1.0.11
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.7
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1.3
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
num2words==0.5.13
numba==0.57.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
omegaconf==2.3.0
openai==1.52.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
openunmix==1.2.1
opt-einsum==3.3.0
optree==0.13.1
orjson==3.8.14
oscrypto==1.3.0
packaging==23.1
paddleocr==2.9.1
paddlepaddle==2.6.2
paddlepaddle-gpu==2.6.2
pandas==2.0.1
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
pdfminer==20191125
pdfminer.six==20221105
peewee==3.17.1
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.2.0
pipeline==0.1.0
platformdirs==3.5.0
preshed==3.0.9
prettytable==3.9.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.23.0
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.21
pycryptodome==3.17
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.15.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pymongo==3.10.0
PyMuPDF==1.22.2
pynput==1.7.7
pyparsing==3.0.9
pypdf==4.3.1
PyPDF2==3.0.1
pypng==0.20220715.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.16.1
pyrsistent==0.19.3
PySocks==1.7.1
PySoundFile==0.9.0.post1
pyspellchecker==0.7.2
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.3
pytk==0.0.2.1
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.4.3
QtPy==2.3.1
RapidFuzz==3.11.0
referencing==0.32.1
regex==2023.5.5
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rlPyCairo==0.2.0
rpds-py==0.17.1
rsa==4.9
ruff==0.9.1
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools-scm==8.0.4
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==6.3.0
smmap==5.0.2
sniffio==1.3.0
snscrape==0.7.0.20230622
sounddevice==0.4.6
soupsieve==2.4.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.0
stack-data==0.6.2
starlette==0.41.3
submitit==1.5.1
svglib==1.5.1
sympy==1.13.1
tablib==3.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
tf-keras==2.15.0
thinc==8.3.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tk==0.1.0
tkinter-tooltip==3.0.0
tokenizers==0.20.0
tomli==2.0.1
tomlkit==0.13.2
toolz==0.12.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.45.1
treetable==0.2.5
twilio==9.2.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uc-micro-py==1.0.2
ultralytics==8.3.69
ultralytics-thop==2.0.14
uri-template==1.2.0
uritools==4.0.3
urllib3==2.0.2
uvicorn==0.22.0
wandb==0.19.6
wasabi==1.1.3
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==4.0.7
word2number==1.1
wrapt==1.14.1
xhtml2pdf==0.2.16
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
yarl==1.9.2
yfinance==0.2.37
